pyproject.toml
[build-system]
    requires = ["setuptools>=42", "wheel"]
    build-backend = "setuptools.build_meta"

[project]
    name = "git-agent"
    version = "0.1.0"
    description = "A Git agent for managing software code and configs on non-k8s servers"
    readme = "README.md"
    requires-python = ">=3.6"
    keywords = ["git", "classical", "infrastructure", "on-premise", "servers"]
    license = { text = "MIT" }
    authors = [{ name = "RS Nikhil Krishna", email = "rsnk96@gmail.com" }]
    classifiers = [
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        "Intended Audience :: Developers",
        "Topic :: Software Development :: Build Tools",
        "Development Status :: 3 - Alpha",
    ]
    dependencies = ["gitpython", "toml"]
    scripts = { "git-agent" = "git_agent.agent:main" }

[tool.setuptools]
    include-package-data = true
    [tool.setuptools.packages]
        find = {}
